requirements.txt
streamlit 

langchain
 langchain-community
  langchain-groq
   langgraph
    python-dotenv
     arxiv
      requests 
      pydantic
       typing-extensions